split lists functions into two files
